👾 refactor submarine systems
